# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        stellarAddress:
          type: string
        score:
          type: number
      required:
        - rank
        - stellarAddress
        - score
...
    FeatureFlag:
      type: object
      properties:
        key:
          type: string
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - key
        - enabled
        - description
        - updatedAt
    FeatureFlagListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FeatureFlag'
      required:
        - data
    FeatureFlagResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/FeatureFlag'
      required:
        - data
    CreateFeatureFlagRequest:
      type: object
      properties:
        key:
          type: string
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
          maxLength: 280
      required:
        - key
        - enabled
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        actor:
          type: string
        targetAddress:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - createdAt
...
paths:
...
  /api/leaderboard:
    get:
      operationId: getLeaderboard
      tags:
        - Leaderboard
      summary: Get global leaderboard
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            maximum: 100
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
        - schema:
            type: boolean
            nullable: true
            default: false
          required: false
          name: refresh
          in: query
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  meta:
                    type: object
                    properties:
                      limit:
                        type: number
                      offset:
                        type: number
                      count:
                        type: number
                      refresh:
                        type: boolean
                    required:
                      - limit
                      - offset
                      - count
                      - refresh
                required:
                  - data
                  - meta
...
  /api/admin/feature-flags:
    get:
      operationId: listAdminFeatureFlags
      tags:
        - Admin
      summary: List configured feature flags
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of feature flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    post:
      operationId: createAdminFeatureFlag
      tags:
        - Admin
      summary: Create a new feature flag
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureFlagRequest'
      responses:
        '201':
          description: Feature flag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Feature flag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin/audit:
    get:
      operationId: getAdminAuditLog
      tags:
        - Admin
      summary: List audit log entries (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: false
          name: action
          in: query
        - schema:
            type: string
          required: false
          name: actor
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: startDate
          in: query
        - schema:
            type: string
            format: date-time
          required: false
          name: endDate
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...